Narrowed conformance criteria, emphasised simplicity and economic rationale
Have narrowed the conformance criteria slightly. We should probably sit down and work out our full set of criteria, with an eye to keeping it short and simple. This will also make it less burdensome for developers to provide feedback, and easier for agencies to run through the full conformance suite. Have also placed a stronger emphasis on 'simplicity', and given the public policy rationale as to why we want to make it easy for developers to use our APIs.